refactor(renderer): extract ContextBar, ModelPicker, panel cards
Second extraction wave for BET-47. Moves three more self-contained
footer/composer widgets out of ChatPanel.tsx:
- ContextBar.tsx: usage bar + stale-cache pill
- ModelPicker.tsx: per-session model dropdown
- PanelCards.tsx: ScheduledTasksCard, WebhooksCard, SecretsCard

ChatPanel.tsx drops from ~5700 to ~4960 LoC. typecheck passes.
