+++ src/renderer/ChatPanel.tsx
@@ -13,7 +13,7 @@
 // No Electron-only deps — only `window.api.*` (the mobile HTTP server will
 // shim that surface).
 
-import { memo, useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from "react";
 import type {
   OpencodeAgent,
   OpencodeCommand,
@@ -24,7 +24,6 @@ import type {
   QuestionRequest,
   ScheduledJob,
   SecretMeta,
-  SecretScope,
   WebhookMeta,
 } from "../shared/types";
 import { useStore } from "./store";
@@ -39,7 +38,6 @@ import type { VoiceAction } from "../shared/types";
 import {
   ASSUMED_CONTEXT_TOKENS,
   formatBytes,
-  ctxStageColor,
   filterCommands,
   dedupeAgainstBuiltins,
   resolveContextLimit,
@@ -73,10 +71,7 @@ import {
   buildTitlePromptInput,
   buildTitleInstruction,
   sanitizeGeneratedTitle,
-  describeCron,
-  describeNextRun,
   type TruncationKind,
-  type ContextBreakdown,
   type StaleCacheResult,
   type PendingDelta,
 } from "./chatUtils";
@@ -96,6 +91,9 @@ import {
 } from "./chatShared";
 import { ActiveTodos, MessageRow, RunningIndicator } from "./MessageRow";
 import { CompactionCard, PermissionCard, QuestionCard, RetryCard } from "./Cards";
+import { ContextBar } from "./ContextBar";
+import { ModelPicker } from "./ModelPicker";
+import { ScheduledTasksCard, SecretsCard, WebhooksCard } from "./PanelCards";
 
 // In-flight attachments tracked alongside the textarea content. Each chip
 // rendered above the input maps to one entry; `status` drives the chip
@@ -4234,751 +4232,6 @@ export function ChatPanel({ sessionId, tmuxSession, windowIndex, cwd, isActive }
   );
 }
 
-
-// ===== Context bar =====
-//
-// Chunky horizontal bar with a dotted "empty" pattern and a SEGMENTED filled
-// portion: fresh-input (paid full rate) | cache.write (warm-up, paid full
-// rate + 25% surcharge) | cache.read (cached, paid ~10%). Color of the
-// fresh segment stages by total usage (green → yellow → orange → red);
-// cache.write uses a steady amber to flag the "this turn cost extra to
-// warm the cache" bucket; cache.read uses a steady muted teal for the
-// "served from cache" bucket.
-//
-// When the session has gone stale (idle past the Anthropic cache TTL the
-// user configured in Settings, AND the cached prefix is non-trivial), an
-// amber `⚠ /clear to save Nk tokens` pill renders to the right of the %.
-// This is the actionable warning: those tokens are about to flip from
-// "served from cache (~10% rate)" to "cache_creation_input_tokens (full
-// rate + surcharge)" on the next user message, so the user can either
-// /clear (save them entirely) or /compact (shrink the prefix first).
-
-// Cache-segment colors — tuned to read as distinct buckets without
-// competing with the stage color of the fresh segment.
-const CACHE_WRITE_COLOR = "#f59e0b"; // amber-500: warm-up, expensive
-const CACHE_READ_COLOR = "#0ea5a4"; // teal-600: cached, cheap
-
-// Format a token count compactly for the inline stale-cache pill
-// ("12k", "120k", "1.2M"). Differs from `formatTokens` (which appends
-// "tokens" and never reaches M-scale) because pill space is tight and
-// we want a millions suffix once a session crosses 1M (Opus 4.7's full
-// window).
-function formatTokensCompact(n: number): string {
-  if (n < 1000) return String(n);
-  if (n < 100_000) return `${(n / 1000).toFixed(1).replace(/\.0$/, "")}k`;
-  if (n < 1_000_000) return `${Math.round(n / 1000)}k`;
-  return `${(n / 1_000_000).toFixed(1).replace(/\.0$/, "")}M`;
-}
-
-function formatIdleDuration(ms: number): string {
-  // Coarse human format for the tooltip — exact precision doesn't help.
-  const totalSec = Math.floor(ms / 1000);
-  const h = Math.floor(totalSec / 3600);
-  const m = Math.floor((totalSec % 3600) / 60);
-  if (h > 0) return `${h}h ${m}m`;
-  if (m > 0) return `${m}m`;
-  return `${totalSec}s`;
-}
-
-function ContextBar({
-  breakdown,
-  limit,
-  staleCache,
-  modelName,
-  tooltip,
-}: {
-  breakdown: ContextBreakdown;
-  limit: number;
-  staleCache: StaleCacheResult;
-  modelName: string | null;
-  tooltip?: string;
-}) {
-  const { pct, segments, freshInput, cacheRead, cacheWrite, totalInput } =
-    breakdown;
-  // Stage color is driven by total usage (so the % digits + fresh slice
-  // share the same warning hue).
-  const fill = ctxStageColor(pct);
-  const dot = `${fill}55`;
-  const segColor = (kind: ContextBreakdown["segments"][number]["kind"]) => {
-    if (kind === "fresh") return fill;
-    if (kind === "cacheWrite") return CACHE_WRITE_COLOR;
-    return CACHE_READ_COLOR;
-  };
-  // Multi-line tooltip showing the full breakdown. Caller appends any
-  // contextual hints (compact recommended, model name, etc.).
-  const breakdownLines = [
-    `${totalInput.toLocaleString()} / ${limit.toLocaleString()} tokens (${pct}%)`,
-    modelName ? `Model window: ${modelName}` : null,
-    cacheRead > 0
-      ? `  · ${cacheRead.toLocaleString()} cache read (cheap, served from cache)`
-      : null,
-    cacheWrite > 0
-      ? `  · ${cacheWrite.toLocaleString()} cache write (warm-up — paid full rate + surcharge this turn)`
-      : null,
-    freshInput > 0
-      ? `  · ${freshInput.toLocaleString()} fresh input (uncached, paid full rate)`
-      : null,
-    tooltip ?? null,
-  ].filter(Boolean);
-  return (
-    <span
-      className="flex items-center gap-1.5 shrink-0"
-      title={breakdownLines.join("\n")}
-    >
-      <span
-        // Keep the `w-24` class — mobile.css selects on it to hide the
-        // track on phones. Don't rename the width without updating that
-        // CSS rule.
-        className="inline-block w-24 h-3 rounded-[2px] overflow-hidden align-middle"
-        style={{
-          backgroundColor: "#1b1e25",
-          backgroundImage: `radial-gradient(circle, ${dot} 1.2px, transparent 1.4px)`,
-          backgroundSize: "4px 4px",
-        }}
-      >
-        {/* Render segments inline; each takes its share of the WHOLE track
-            (not of the filled portion), so widths sum to `pct` and the
-            empty remainder is the dotted pattern beneath. */}
-        {segments.map((s, i) =>
-          s.pct > 0 ? (
-            <span
-              key={s.kind}
-              className="inline-block h-full align-top"
-              style={{
-                width: `${s.pct}%`,
-                backgroundColor: segColor(s.kind),
-                // Subtle inset between segments so adjacent slices read as
-                // distinct buckets even when their colors are close.
-                boxShadow:
-                  i > 0 ? "inset 1px 0 0 rgba(0,0,0,0.35)" : undefined,
-              }}
-            />
-          ) : null,
-        )}
-      </span>
-      <span
-        className="tabular-nums text-[12px] font-semibold"
-        style={{ color: fill }}
-      >
-        {pct}%
-      </span>
-      {staleCache.isStale && (
-        // Visible stale-cache pill — only appears when the session has
-        // been idle past the configured Anthropic cache TTL (5m or 1h,
-        // see Settings) AND the cached prefix is non-trivial. The next
-        // user message will pay full rate + surcharge to re-warm exactly
-        // these tokens; /clear avoids the bill entirely.
-        <span
-          className="tabular-nums text-[11px] font-medium px-1.5 rounded-sm shrink-0"
-          style={{
-            color: CACHE_WRITE_COLOR,
-            backgroundColor: `${CACHE_WRITE_COLOR}1f`,
-          }}
-          title={[
-            `Session idle for ${formatIdleDuration(staleCache.idleMs)} — prompt cache has expired.`,
-            `${staleCache.staleTokens.toLocaleString()} tokens currently in cache will be re-billed as cache_creation_input_tokens on your next message (full input rate + 25% surcharge, or 2× for 1h cache).`,
-            "",
-            "Actions:",
-            "  · /clear  — start a fresh session, skip the re-warm cost entirely",
-            "  · /compact — shrink the prefix before re-warming",
-            "",
-            "(Cache TTL is set by opencode; bui predicts staleness from the Settings → Prompt cache TTL value. If this fires at the wrong time, that setting probably doesn't match opencode's cache_control.ttl.)",
-          ].join("\n")}
-        >
-          ⚠ /clear to save {formatTokensCompact(staleCache.staleTokens)} tokens
-        </span>
-      )}
-    </span>
-  );
-}
-
-// ===== Model picker =====
-//
-// Compact dropdown that shows the active model on the left (either the
-// user-selected override or the last model used by the server). Clicking the
-// label expands an absolutely-positioned list above the footer. Selecting a
-// row sets the per-session override (persisted in localStorage by ChatPanel);
-// the "Default" row clears it so prompt_async falls back to opencode's default.
-
-function ModelPicker({
-  modelLabel,
-  models,
-  modelOverride,
-  defaultModel,
-  onOpen,
-  onSelect,
-}: {
-  modelLabel: string | null;
-  models: OpencodeModel[] | null;
-  modelOverride: ModelSelection | null;
-  defaultModel: { providerID: string; modelID: string } | null;
-  onOpen: () => void;
-  onSelect: (m: ModelSelection | null) => void;
-}) {
-  const [open, setOpen] = useState(false);
-  const rootRef = useRef<HTMLDivElement>(null);
-
-  // Click-away to dismiss the dropdown. Using mousedown (not click) so we
-  // close before the inner button's onClick re-toggles. Buttons inside the
-  // popup still fire their onClick because we check containment.
-  useEffect(() => {
-    if (!open) return;
-    const onDown = (e: MouseEvent) => {
-      if (rootRef.current && !rootRef.current.contains(e.target as Node)) {
-        setOpen(false);
-      }
-    };
-    document.addEventListener("mousedown", onDown);
-    return () => document.removeEventListener("mousedown", onDown);
-  }, [open]);
-
-  // Label precedence — show what will be used for the NEXT prompt, not what
-  // the last response happened to use:
-  //   1. user-picked override (explicit choice)
-  //   2. server default for the first connected provider (when no override —
-  //      "Server default" was picked, so we show the actual default name)
-  //   3. last assistant message's modelID (fallback while defaultModel still loading)
-  //   4. "opencode" stub (initial render, nothing loaded yet)
-  const label = modelOverride
-    ? `${modelOverride.providerID}/${modelOverride.modelID}${modelOverride.variant ? `@${modelOverride.variant}` : ""}`
-    : defaultModel
-      ? `${defaultModel.providerID}/${defaultModel.modelID}`
-      : modelLabel
-        ? modelLabel
-        : null;
-
-  // Group models by providerID so the list reads e.g. "anthropic" → 3 models.
-  const groups = useMemo(() => {
-    if (!models) return null;
-    const map = new Map<string, OpencodeModel[]>();
-    for (const m of models) {
-      if (m.enabled === false || m.status === "deprecated") continue;
-      const arr = map.get(m.providerID) ?? [];
-      arr.push(m);
-      map.set(m.providerID, arr);
-    }
-    return [...map.entries()].sort(([a], [b]) => a.localeCompare(b));
-  }, [models]);
-
-  const isActive = (m: OpencodeModel, variantId?: string): boolean => {
-    if (modelOverride) {
-      return (
-        modelOverride.providerID === m.providerID &&
-        modelOverride.modelID === m.id &&
-        (modelOverride.variant ?? undefined) === variantId
-      );
-    }
-    return false;
-  };
-
-  return (
-    <div ref={rootRef} className="relative min-w-0">
-      <button
-        className="truncate text-[12px] text-text-muted hover:text-text flex items-center gap-1"
-        onClick={() => {
-          if (!open) onOpen();
-          setOpen((v) => !v);
-        }}
-        title="Pick model for next prompt"
-      >
-        <span className="truncate">{label ?? <span className="opacity-60">opencode</span>}</span>
-        <span className="text-text-faint text-[9px]">▾</span>
-      </button>
-      {open && (
-        <div
-          className="absolute left-0 bottom-full mb-1 z-20 min-w-[240px] max-h-[360px] overflow-y-auto rounded border border-border bg-bg-elev shadow-lg text-[12px]"
-        >
-          <button
-            onClick={() => {
-              onSelect(null);
-              setOpen(false);
-            }}
-            className={
-              "w-full text-left px-2 py-1 hover:bg-bg-soft border-b border-border " +
-              (modelOverride == null ? "text-text" : "text-text-muted")
-            }
-          >
-            <span className="mr-1" style={{ color: modelOverride == null ? CLAUDE_ORANGE : "transparent" }}>●</span>
-            Server default
-          </button>
-          {!groups && (
-            <div className="px-2 py-2 text-text-faint">Loading…</div>
-          )}
-          {groups?.length === 0 && (
-            <div className="px-2 py-2 text-text-faint">No models</div>
-          )}
-          {groups?.map(([providerID, ms]) => (
-            <div key={providerID} className="py-1">
-              <div className="px-2 py-0.5 text-[10px] uppercase tracking-wider text-text-faint">
-                {providerID}
-              </div>
-              {ms.map((m) => {
-                const variants = m.variants ?? [];
-                return (
-                  <div key={m.id}>
-                    <button
-                      onClick={() => {
-                        onSelect({ providerID: m.providerID, modelID: m.id });
-                        setOpen(false);
-                      }}
-                      className={
-                        "w-full text-left px-2 py-0.5 hover:bg-bg-soft flex justify-between gap-2 " +
-                        (isActive(m) ? "text-text" : "text-text-muted")
-                      }
-                    >
-                      <span className="truncate flex items-center gap-1">
-                        <span style={{ color: isActive(m) ? CLAUDE_ORANGE : "transparent" }}>●</span>
-                        <span>{m.name}</span>
-                      </span>
-                      {m.limit?.context ? (
-                        <span className="text-text-faint text-[10px] shrink-0">
-                          {Math.round(m.limit.context / 1000)}k
-                        </span>
-                      ) : null}
-                    </button>
-                    {variants.map((v) => (
-                      <button
-                        key={v.id}
-                        onClick={() => {
-                          onSelect({ providerID: m.providerID, modelID: m.id, variant: v.id });
-                          setOpen(false);
-                        }}
-                        className={
-                          "w-full text-left pl-6 pr-2 py-0.5 hover:bg-bg-soft text-[11px] " +
-                          (isActive(m, v.id) ? "text-text" : "text-text-faint")
-                        }
-                      >
-                        <span style={{ color: isActive(m, v.id) ? CLAUDE_ORANGE : "transparent" }}>●</span>{" "}
-                        @{v.id}
-                      </button>
-                    ))}
-                  </div>
-                );
-              })}
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ===== Session toolbar =====
-//
-// Three compact buttons in the footer right rail: fork (copy current state
-// into a new session+window), compact (in-place summarization to free
-// context), delete (DELETE on the server + kill the tmux window). Fork and
-// delete are disabled when the panel doesn't know its owning tmux window.
-
-// ScheduledTasksCard — pinned card above the composer showing this session's
-// scheduled prompts (created by the AI's `schedule` opencode tool) with a
-// per-row delete. A card (not a footer item) so it renders on BOTH desktop and
-// mobile with no mobile-CSS edits. See docs/bui-tools-scheduler.md.
-const ScheduledTasksCard = memo(function ScheduledTasksCard({
-  jobs,
-  error,
-  onDelete,
-  onClose,
-}: {
-  jobs: ScheduledJob[];
-  error: string | null;
-  onDelete: (id: string) => void;
-  onClose: () => void;
-}) {
-  // Click-outside-to-dismiss: the close (×) button sits directly above the
-  // first row's cancel button, so a mis-tap on the close used to delete a job.
-  // Dismissing by clicking anywhere outside the card removes that hazard.
-  const cardRef = useRef<HTMLDivElement | null>(null);
-  useEffect(() => {
-    const onDown = (e: MouseEvent) => {
-      if (cardRef.current && !cardRef.current.contains(e.target as Node)) {
-        onClose();
-      }
-    };
-    // Defer registration to the next tick so the same click that opened the
-    // card (from the toolbar button) doesn't immediately close it.
-    const t = setTimeout(() => document.addEventListener("mousedown", onDown), 0);
-    return () => {
-      clearTimeout(t);
-      document.removeEventListener("mousedown", onDown);
-    };
-  }, [onClose]);
-  return (
-    <div
-      ref={cardRef}
-      className="rounded-md border bg-bg-elev px-3 py-2 text-[12px]"
-      style={{ borderColor: CLAUDE_ORANGE + "55" }}
-    >
-      <div className="flex items-center gap-2 mb-1">
-        <span style={{ color: CLAUDE_ORANGE }}>⏰</span>
-        <span className="text-text">Scheduled</span>
-        {jobs.length > 0 && <span className="text-text-faint">· {jobs.length}</span>}
-        <button
-          onClick={onClose}
-          className="ml-auto px-1.5 rounded text-text-faint hover:text-text-muted text-[14px]"
-          title="Close (or click outside)"
-        >
-          ×
-        </button>
-      </div>
-      {error ? (
-        <div className="text-red-400 break-words">{error}</div>
-      ) : jobs.length === 0 ? (
-        <div className="text-text-muted">No scheduled tasks in this session.</div>
-      ) : (
-        <div className="flex flex-col gap-1.5">
-          {jobs.map((j) => {
-            const next = describeNextRun(j.cron, j.recurring);
-            return (
-              <div key={j.id} className="flex items-center gap-2">
-                <div className="min-w-0 flex-1">
-                  <div className="text-text truncate" title={j.prompt}>
-                    {j.label || j.prompt}
-                  </div>
-                  <div className="flex items-center gap-2 text-text-faint font-mono text-[11px]">
-                    <span className="shrink-0">
-                      {describeCron(j.cron, j.recurring)}
-                    </span>
-                    {next && (
-                      <span
-                        className="shrink-0 truncate"
-                        title="Next run"
-                        style={{ color: CLAUDE_ORANGE }}
-                      >
-                        · next {next}
-                      </span>
-                    )}
-                  </div>
-                </div>
-                <button
-                  onClick={() => onDelete(j.id)}
-                  className="shrink-0 px-2 py-0.5 rounded text-red-400 hover:bg-red-500/10 border border-red-500/30 text-[11px]"
-                  title="Cancel this scheduled task"
-                >
-                  Cancel
-                </button>
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-});
-
-// WebhooksCard — pinned card above the composer listing this session's inbound
-// webhooks (created by the AI's `webhook` opencode tool) with a per-row revoke.
-// A card (not a footer item) so it renders on BOTH desktop and mobile with no
-// mobile-CSS edits. List is metadata only — the signing secret is shown once at
-// create (in the agent's tool result) and never re-exposed here.
-// See docs/bui-tools-webhook.md.
-const WebhooksCard = memo(function WebhooksCard({
-  hooks,
-  error,
-  onDelete,
-  onClose,
-}: {
-  hooks: WebhookMeta[];
-  error: string | null;
-  onDelete: (id: string) => void;
-  onClose: () => void;
-}) {
-  const cardRef = useRef<HTMLDivElement | null>(null);
-  const [copied, setCopied] = useState<string | null>(null);
-  useEffect(() => {
-    const onDown = (e: MouseEvent) => {
-      if (cardRef.current && !cardRef.current.contains(e.target as Node)) onClose();
-    };
-    const t = setTimeout(() => document.addEventListener("mousedown", onDown), 0);
-    return () => {
-      clearTimeout(t);
-      document.removeEventListener("mousedown", onDown);
-    };
-  }, [onClose]);
-  const copyUrl = (url: string, id: string) => {
-    void navigator.clipboard?.writeText(url).then(
-      () => {
-        setCopied(id);
-        setTimeout(() => setCopied((c) => (c === id ? null : c)), 1500);
-      },
-      () => { /* clipboard blocked — no-op */ },
-    );
-  };
-  return (
-    <div
-      ref={cardRef}
-      className="rounded-md border bg-bg-elev px-3 py-2 text-[12px]"
-      style={{ borderColor: CLAUDE_ORANGE + "55" }}
-    >
-      <div className="flex items-center gap-2 mb-1">
-        <span style={{ color: CLAUDE_ORANGE }}>🪝</span>
-        <span className="text-text">Webhooks</span>
-        {hooks.length > 0 && <span className="text-text-faint">· {hooks.length}</span>}
-        <button
-          onClick={onClose}
-          className="ml-auto px-1.5 rounded text-text-faint hover:text-text-muted text-[14px]"
-          title="Close (or click outside)"
-        >
-          ×
-        </button>
-      </div>
-      {error ? (
-        <div className="text-red-400 break-words">{error}</div>
-      ) : hooks.length === 0 ? (
-        <div className="text-text-muted">
-          No webhooks in this session. Ask the agent to create one (e.g. “have
-          Multica ping this session when the task finishes”).
-        </div>
-      ) : (
-        <div className="flex flex-col gap-1.5">
-          {hooks.map((h) => {
-            const last =
-              h.lastDeliveredAt != null
-                ? `${new Date(h.lastDeliveredAt).toLocaleString()} · ${h.deliveries}×`
-                : "never fired";
-            return (
-              <div key={h.id} className="flex items-center gap-2">
-                <div className="min-w-0 flex-1">
-                  <div className="flex items-center gap-1.5">
-                    <span className="text-text truncate" title={h.label}>
-                      {h.label}
-                    </span>
-                    {h.unsigned && (
-                      <span
-                        className="shrink-0 px-1 rounded text-red-400 border border-red-500/30 text-[10px]"
-                        title="No signature required — anyone with the URL can trigger this hook"
-                      >
-                        unsigned
-                      </span>
-                    )}
-                  </div>
-                  <div className="flex items-center gap-2 text-text-faint font-mono text-[11px]">
-                    {h.url && (
-                      <button
-                        onClick={() => copyUrl(h.url as string, h.id)}
-                        className="shrink-0 truncate max-w-[200px] hover:text-text-muted underline decoration-dotted"
-                        title={`Copy delivery URL\n${h.url}`}
-                      >
-                        {copied === h.id ? "copied!" : h.url.replace(/^https?:\/\//, "")}
-                      </button>
-                    )}
-                    <span className="shrink-0 truncate" title="Last delivery">
-                      · {last}
-                    </span>
-                  </div>
-                </div>
-                <button
-                  onClick={() => onDelete(h.id)}
-                  className="shrink-0 px-2 py-0.5 rounded text-red-400 hover:bg-red-500/10 border border-red-500/30 text-[11px]"
-                  title="Revoke this webhook (further POSTs will 404)"
-                >
-                  Revoke
-                </button>
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-});
-
-// SecretsCard — pinned card above the composer for managing the secrets the
-// agent can use. The user types a key + value here; the value travels to the
-// box (renderer → IPC/RPC → bui-server store) and is NEVER returned or shown
-// again — the list is metadata only (key, scope, hint). Agents read secrets via
-// the secret_list / secret_provide opencode tools, which materialize the value
-// to a 0600 file on the box and hand the agent only the path, so the value
-// never enters the AI transcript. Modeled on ScheduledTasksCard so it renders
-// on desktop AND mobile with no mobile-CSS edits.
-const SecretsCard = memo(function SecretsCard({
-  secrets,
-  error,
-  sessionId,
-  onSave,
-  onDelete,
-  onClose,
-}: {
-  secrets: SecretMeta[];
-  error: string | null;
-  sessionId: string;
-  onSave: (input: {
-    key: string;
-    value: string;
-    scope: SecretScope;
-    sessionID?: string | null;
-    hint?: string;
-  }) => Promise<boolean>;
-  onDelete: (id: string) => void;
-  onClose: () => void;
-}) {
-  const [key, setKey] = useState("");
-  const [value, setValue] = useState("");
-  const [scope, setScope] = useState<SecretScope>("shared");
-  const [hint, setHint] = useState("");
-  const [saving, setSaving] = useState(false);
-
-  const keyValid = /^[A-Za-z_][A-Za-z0-9_]{0,63}$/.test(key);
-  const canSave = keyValid && value.length > 0 && !saving;
-
-  const submit = () => {
-    if (!canSave) return;
-    setSaving(true);
-    void onSave({
-      key,
-      value,
-      scope,
-      // Pass sessionID for session scope (the owner) AND project scope (so the
-      // server resolves the workspace name from this chat's session).
-      sessionID: scope === "session" || scope === "project" ? sessionId : null,
-      hint: hint.trim() || undefined,
-    }).then((ok) => {
-      setSaving(false);
-      if (ok) {
-        // Clear value immediately (don't keep the secret in component state),
-        // and reset the form for the next entry.
-        setKey("");
-        setValue("");
-        setHint("");
-      }
-    });
-  };
-
-  return (
-    <div
-      className="rounded-md border bg-bg-elev px-3 py-2 text-[12px]"
-      style={{ borderColor: CLAUDE_ORANGE + "55" }}
-    >
-      <div className="flex items-center gap-2 mb-1.5">
-        <span style={{ color: CLAUDE_ORANGE }}>🔑</span>
-        <span className="text-text">Secrets</span>
-        {secrets.length > 0 && <span className="text-text-faint">· {secrets.length}</span>}
-        <button
-          onClick={onClose}
-          className="ml-auto px-1 rounded text-text-faint hover:text-text-muted"
-          title="Close"
-        >
-          ×
-        </button>
-      </div>
-
-      {/* Add / update form */}
-      <div className="flex flex-col gap-1.5 mb-2">
-        <div className="flex flex-wrap gap-1.5">
-          <input
-            value={key}
-            onChange={(e) => setKey(e.target.value)}
-            placeholder="KEY (e.g. GITHUB_PAT)"
-            spellCheck={false}
-            autoCapitalize="off"
-            autoCorrect="off"
-            className={`min-w-0 flex-1 rounded border bg-bg px-1.5 py-1 font-mono text-text outline-none ${
-              key && !keyValid ? "border-red-500/60" : "border-border"
-            }`}
-          />
-          <select
-            value={scope}
-            onChange={(e) => setScope(e.target.value as SecretScope)}
-            className="rounded border border-border bg-bg px-1.5 py-1 text-text outline-none"
-            title="shared = every session · project = every chat in this workspace · session = only this chat"
-          >
-            <option value="shared">shared</option>
-            <option value="project">this project</option>
-            <option value="session">this session</option>
-          </select>
-        </div>
-        <input
-          value={value}
-          onChange={(e) => setValue(e.target.value)}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") {
-              e.preventDefault();
-              submit();
-            }
-          }}
-          placeholder="value (stored on the box; never shown again)"
-          type="password"
-          spellCheck={false}
-          autoCapitalize="off"
-          autoCorrect="off"
-          className="w-full rounded border border-border bg-bg px-1.5 py-1 font-mono text-text outline-none"
-        />
-        <div className="flex flex-wrap gap-1.5">
-          <input
-            value={hint}
-            onChange={(e) => setHint(e.target.value)}
-            placeholder="hint for the agent (optional, e.g. 'git push token')"
-            className="min-w-0 flex-1 rounded border border-border bg-bg px-1.5 py-1 text-text outline-none"
-          />
-          <button
-            onClick={submit}
-            disabled={!canSave}
-            className="shrink-0 px-2 py-1 rounded border disabled:opacity-40"
-            style={{ borderColor: CLAUDE_ORANGE + "88", color: CLAUDE_ORANGE }}
-            title="Store this secret on the box"
-          >
-            {saving ? "saving…" : "Save"}
-          </button>
-        </div>
-        {key && !keyValid && (
-          <div className="text-red-400 text-[11px]">
-            Key must start with a letter/underscore, then letters/digits/underscores (max 64).
-          </div>
-        )}
-      </div>
-
-      {error && <div className="text-red-400 break-words mb-1">{error}</div>}
-
-      {/* Existing secrets (metadata only — no values) */}
-      {secrets.length === 0 ? (
-        <div className="text-text-muted">
-          No secrets yet. Add one above; the agent uses it via the secret_provide tool
-          without ever seeing the value.
-        </div>
-      ) : (
-        <div className="flex flex-col gap-1.5 border-t border-border pt-1.5">
-          {secrets.map((s) => (
-            <div key={s.id} className="flex items-center gap-2">
-              <div className="min-w-0 flex-1">
-                <div className="flex items-center gap-1.5">
-                  <span className="text-text font-mono truncate">{s.key}</span>
-                  <span
-                    className="shrink-0 rounded px-1 text-[10px] text-text-faint border border-border"
-                    title={
-                      s.scope === "shared"
-                        ? "Available to every session"
-                        : s.scope === "project"
-                          ? `Available to every chat in project "${s.project ?? ""}"`
-                          : "Available only to this session"
-                    }
-                  >
-                    {s.scope === "shared"
-                      ? "shared"
-                      : s.scope === "project"
-                        ? `project:${s.project ?? "?"}`
-                        : "session"}
-                  </span>
-                </div>
-                {s.hint && (
-                  <div className="text-text-faint text-[11px] truncate" title={s.hint}>
-                    {s.hint}
-                  </div>
-                )}
-              </div>
-              <button
-                onClick={() => onDelete(s.id)}
-                className="shrink-0 px-1.5 py-0.5 rounded text-red-400 hover:bg-red-500/10 border border-red-500/30"
-                title="Delete this secret"
-              >
-                ✕
-              </button>
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-});
 
 // SessionToolbar — footer affordances. fork / compact / delete moved out of the
 // footer (they live in the header ⋯ menu); only the ⏰ schedules toggle remains
